Separate medium container into a concern
