+++ src/api/.rubocop_todo.yml
@@ -296,6 +296,7 @@ Metrics/CyclomaticComplexity:
     - 'app/models/kiwi/image/xml_parser.rb'
     - 'app/models/owner_search/assignee.rb'
     - 'app/models/package.rb'
+    - 'app/models/concerns/package_medium_container.rb'
     - 'app/models/package_build_status.rb'
     - 'app/models/patchinfo.rb'
     - 'app/models/product.rb'
@@ -410,6 +411,7 @@ Metrics/PerceivedComplexity:
     - 'app/models/kiwi/image/xml_parser.rb'
     - 'app/models/owner_search/assignee.rb'
     - 'app/models/package.rb'
+    - 'app/models/concerns/package_medium_container.rb'
     - 'app/models/package_build_status.rb'
     - 'app/models/patchinfo.rb'
     - 'app/models/product.rb'
